requirements.txt
gradio>=4.0.0
requests>=2.28.0
python-dotenv>=0.19.0
pydantic>=1.10.0
librosa>=0.9.2
numpy>=1.21.0
